Separate out status and progress ropes
