+++ packages/ui/src/components/progress-rope/components/base-rope-group-step/base-rope-group-step.utils.ts
