+++ packages/ui/src/components/progress-rope/components/base-rope-group-step/index.ts
@@ -1,0 +1,2 @@
+export { BaseRopeGroupStep } from './base-rope-group-step.component.js';
+export { type BaseRopeGroupStepProps } from './base-rope-group-step.types.js';
