+++ packages/ui/src/components/progress-rope/components/base-rope-step/base-rope-step.styles.ts
@@ -1,11 +1,11 @@
 import { tv } from 'tailwind-variants';
 
-export const styles = tv({
+export const baseRopeStepStyles = tv({
   slots: {
     base: 'relative flex w-full gap-2 pt-[0.5rem] pb-[0.875rem] typography-body-10 leading-loose outline-transparent transition-colors',
     circle: 'relative z-10 bg-background-white transition-colors',
     label: 'grid w-full grid-cols-[auto_1fr] items-center',
-    subText: 'col-start-2 pt-[0.75rem] typography-body-10 font-normal text-text-muted',
+    subText: 'col-start-2 pt-1 typography-body-10 font-normal text-text-muted',
   },
   variants: {
     variant: {
