+++ packages/ui/src/components/progress-rope/components/base-rope/base-rope.styles.ts
@@ -1,5 +1,5 @@
 import { tv } from 'tailwind-variants';
 
-export const styles = tv({
+export const baseRopeStyles = tv({
   base: 'relative after:absolute after:top-[1.0625rem] after:bottom-4 after:left-1 after:z-0 after:block after:border-l-2 after:border-dotted after:border-border-muted-strong',
 });
