+++ packages/ui/src/components/progress-rope/components/index.ts
@@ -1,2 +1,2 @@
 export * from './progress-rope-step/index.js';
-export * from './progress-rope-group-step/index.js';
+export * from './base-rope-group-step/index.js';
