--- packages/ui/src/components/progress-rope/components/progress-rope-group-step/index.ts
@@ -1,2 +1,0 @@
-export { ProgressRopeGroupStep } from './progress-rope-group-step.component.js';
-export { type ProgressRopeGroupStepProps } from './progress-rope-group-step.types.js';
