+++ packages/ui/src/components/progress-rope/components/status-rope-step/status-rope-step.types.ts
@@ -1,0 +1,14 @@
+import { ReactNode } from 'react';
+
+import { RopeStepProps } from '../base-rope-step/base-rope-step.types.js';
+
+export type StatusRopeStepProps = RopeStepProps & {
+  /**
+   * Additional information associated with the step
+   */
+  subText: ReactNode;
+  /**
+   * Step text
+   */
+  text: ReactNode;
+};
