+++ packages/ui/src/components/progress-rope/index.ts
@@ -1,3 +1,4 @@
 export { ProgressRope } from './progress-rope.component.js';
 export { StatusRope } from './status-rope.component.js';
-export { type ProgressRopeProps, type StatusRopeProps } from './progress-rope.types.js';
+export { type ProgressRopeProps } from './progress-rope.types.js';
+export { type StatusRopeProps } from './status-rope.types.js';
